pyproject.toml
[project]
name = "LexiSight"
version = "0.1.0"
description = "OCR Microservice"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "python-multipart",
    "pydantic>=2.0.0",
    "pydantic-settings",
    "pdf2image",
    "python-dotenv",
    "pillow",
    # vllm is provided by the base image in Docker, 
    # but listed here for local dev if on Linux/WSL
    "vllm>=0.6.0; sys_platform != 'win32'", 
    "transformers>=4.46.0",
    "accelerate>=0.26.0",
    "qwen-vl-utils"
]

[tool.uv]
package = false